Adding site-specifc-master fuctionality. This allows a site to modify the look of the entire site by adding a tcl/adp pair named "site-specific-master" and setting the "UseSiteSpecificMaster" ad_param of the parent acs-subsite package. The param defaults to "false".
Why not just edit the "default-master" files? Well, if you break
default master much of the site will be broken. This feature allows
you to experiment with the site-specific-master with the knoledge that
with the flip of a parameter the pristine "default-master" will be
used. And there's no risk for committers to check in their
custom site-specific-master files into the tree unlike with editing
def-master.