Part of this commit is fixing #1176 (just a few acs-lang strings and properly splitting the file into *-proc.tcl and *-init.tcl part).  Part of this fix (catalog file) was accidentally commited before by Joel
The bigger part is changing load order of library files so that
acs-lang-init.tcl gets loaded before all other *-init.tcl files.

This is necessary because acs-lang is so fundamental and it is quite possible
(as we have seen with the example of date-init.tcl) that there are *-init.tcl
files that need translated strings.