Provide developer with the option to declare compound resources (CSS or JavaScript) to optimize performance or to provide customized themes. New public function:
    template::head::includes -container .. -parts ...

if an included resource is encountered during rendering of the page,
it is ignored. The compound resource management is aligned resource
flushing (when a compound resource is flushed, it stop shadowing
potentially other resources. An additional option could be to
add an "-parts" flag to add_link/add_script.

added simple regression test for basic functionality