Extend API to allow registration for URNs.
URNs provide an abstraction and a single place for e.g updating
references to external resources when switching between a CDN and a
local resource, or when a resource should be updated. Instead of
adding e.g. a CDN url via tempate::head::add_script, one can add an
URN and control its content from a single place.

Example:
   template::register_urn \
       -urn urn:ad:js:jquery \
       -resource /resources/xowiki/jquery/jquery.min.js