apm reloading reform (part 1)
Blueprint reloading updates the blueprint of nsd, which has the
consequence the also threads for running scheduled procedures can be updated.
So far blueprint reloading is just tested with NaviServer, but should work with
AOLserver as well (modulo bugs).

To switch between the classic and blueprint reloading, edit acs-tcl/tcl/apm-procs.tcl
and set ::apm::reloading to "classic" or "blueprint".

Some functions were refactored to ease this transition.