Declass constraint names breaking the very strict OpenACS naming convention from errors to warnings
Rationale: it is sometimes impossible to have constraint names that are both < 30 chars and consistent with the naming scheme 'table_name'_'column_name'_'constraint_type' for all those defined on a particular table (see e.g. acs_named_objects).
Breaking this convention might be suboptimal, but is not harmful in itself.