ad_sign: new optional parameter "user_binding"
The parameter user_binding allows to bind a signature to a user.
When the value is "-1" only the user who created the signature can
obtain the value again. A value of 0 (default) means no user binding.
The permissible values might be extended in the future.

bump version number to 5.10.0d24