auth::authority::get, auth::authority::get_element: allow to omit authority_id
Now, these functions can be used to get data from the default authority. In furture
versions, this function should deliver the right default authority for the
a register subsite. Also, get_authority_options should return the available options
for a register subsite.

Currently, the logic is different in various scripts...