Add xml-rpc system.* methods as exceptions in the naming__proc_naming test.
The xml-rpc package defines procs of the form system.*, that not only pollute the root namespace, but also contain the '.' character, which is not allowed by OpenACS naming standards.

The package is designed in a way that XML-RPC methods are implemented as 'ad_procs' of the same name (see package doc), and the system.* methods are standard XML-RPC methods (see XML-RPC spec). One could also expect current users of the package to have their own local 'whatever.verb' methods/ad_procs in the root namespace as well. Until future refactoring of the package to decouple XML-RPC method and proc names is done, this exception is added.