fix for issue #3445
a broken query (without a table alias) was
in the .xql file, where the correct query was
in the source (where it was ignored). This
can be quite confusing. It is probably
a good idea to improve code locality
and use .xql just in cases, where the
query is NOT included in the source.
I took the opportunity to fix more such cases
in this file.

Many thanks to Tony Kirkham for the bug report.


 CVS: ----------------------------------------------------------------------