added "ad_conn bot_p" to check, whether request was initiated by a bot
This feature is based on a simple heuristic based on the user-agent
(which can be certainly extended). It is useful to avoid e.g. Google
bot to run into "notifications subscribe" + login attempts, which
are useful for not-logged-in user, but not for bots. These attemps
lead to failures in google statistics that might reduce the google
ranking of a web site.