Fix for issue #3446
This fixes a bug, where the ordering of the site-nodes in the answer
set of the select_child_site_nodes query in site_node::update_cache
led to disappearing entries in the site map, when additional
items were added. This bug happened in the old-style
(nsv-based) site-node impementation, which is used, when no
xotcl is available.

Many thanks to Keith Paskett for debugging and the fix.