Add a second period to new lines starting with one ('dot-stuffing'), as defined by RFC 5321 section 4.5.2, before sending mail via nssmptd.
According to the RFC:
- "Before sending a line of mail text, the SMTP client checks the
first character of the line.  If it is a period, one additional period is inserted at
the beginning of the line."

- "When a line of mail text is received by the SMTP server, it checks the line.  If the
line is composed of a single period, it is treated as the end of mail indicator.  If
the first character is a period and there are other characters on the line, the first
character is deleted."

This change adds the second period at the beginning of a line suggested in the first
paragraph, preventing nssmptd from deleting dots by acting as mentioned in the second one.

In the case of lines comprised of just a single period, '::mime::qp_encode' already takes
care of them by converting the '.' into '=2E', so they are not affected by this.