avoid burst cache hits by using nsvs instead of ns_cache for secret token caching
The problem with the prevision version was that ns_cache has no equivalent
to nsv_array to get in an atomic fashion all key value pairs from the cache.
thefore, for the per-thread caching, a tight loop was used processing
the result of ns_cache_keys. If this was necessary in a high number of
threads, higher lock times were possible (~100ms). The new version
should not face this problem.