Revert to previous strategy of determining available subcommands. Running the provided subcommand is dangerous, since the command might have side effects, might require connections, etc.
The feature test for "ns_http run" was incorrect, since it was not
introduced with 4.99.15 (as the comment suggested), but already with
4.99.4 (2017)... but at this time, the command did not return a dict,
which is assumed in the http-client procs. One can use "ns_http stats"
as an indicator for this feature, which was introduced at the time
when "ns_http run" was starting to return dicts as results.

Provide a more precise name for the feature test of ns_http returning
dicts (naming the feature simply "ns_http" is not a good idea, since
the command could of course be used already before the feature change).