fixed bug: exception in test case auth_use_email_for_login_p caused test to terminate
By adding an "ad_try", we can see that the intention to create
a user with no username failed. Probably this was working at some time,
not sure, if it still should.

This is could be a general problem with "aa_run_with_teardown" to
terminate too early in such cases.