Added page_contract filter "object_id"
This change adds the page_contract filter "object_id", which validates
values whether these are syntactically acceptable as object_ids in
PostgreSQL and Oracle.

Note that before one is able to use the filter, the server has to be
restarted.  Otherwise, when e.g. "apm/version_reload.tcl" would be
executed with the new filter, it would fail. So, one has to be careful
on update scripts, when people upgrade from old version not to create
a blocking mutual blocking condition.