Fixed a bug in the request processor, when URL is /%3F
The problem was that /%3F corresponds to a URL which is literally '/?'
(question mark is not the separator for query variables). In this case
a "string match" operation to determine the suffix based on this
string will lead to unexpected characters since '?' is a match
character. This lead in turn to a problem with redirects to the
internally redirect of custom error pages. So, in this case (and
probably others) the custom error page was not displayed.