Notification request cleanup refactoring
One of the most expensive operations in large site is the cleanup for
notification_requests in situations, where the user has lost
permissions on an object, on which the user wanted to receive
notifications.  This check was performed previously in
notification::sweep::cleanup_notifications via a permission check over
all notification requests, which can be very costly on large
sites. This change moves this cleanup into the actual notification
sending, where the permissions have to be sent anyhow.