Change caching of driver keys
There seems to be a bug in "db_driverkey", when OpenACS connects to
multiple databases [1]. The situation with "db_driverkey" is
complicated by the fact that this function is sometimes called with a
database name in its argument "dbn", and sometimes with a database
handle. This is something to look at in the future.

The new version removes the per-thread caching, which was introduced
to reduce the high number of mutex locks caused in busy machines. The
new version is better by a factor of 2 than older versions, but relies
on NaviServer.

[1] https://openacs.org/forums/message-view?message_id=5537865