Provide a default timeout of 5m to all ns_cache besed caches created via the acs-cache infrastructure
Background: in case, a db entry is to be cached, where the db query
does not finish (e.g. runs into a deadlock), subsequent attempts for
the same value will be blocked as well by the mutex protecting the
entry to be cached. Therefore, later requests to this entry might queue
up in the server. By providing a timeout, this mutex will terminate
latest when the timeout is reached.

The default value for the timeout is set to 5m. In case a certain
cache might issue requests running longer than this, the cache has
to be configured with a larger timeout value.