Added per-request-cache for parameter::get similar to parameter handling in xotcl-core
This change improves scalability especially on sites with high numbers
of concurrent requests by reducing the number of required locks in
general. This change keeps the parameter dict in a per-request cache
to reduce potentially high number of nsv locks, when parameters of a
package are queried a often per request. Without this change we see on
some sites > 100 locks on the ad_param-nsv per request. Alone the
request-procesor queries a larger set of parameters from the acs-core
package, which requires now a single lock operation.

Using the per-request cache has the advantave of avoiding locks and
keeping the internal representations of tcl-objs (whereas ns_cache or
nsvs use plain strings).