Fixes for Oracle 19c: quick port of missing tables in acs-mail-lite from PostgreSQL
This is q quick port of the table available from PostgreSQL, which
assures that a new instance of OpenACS on Oracle can run without
spitting out frequent error messages.  In essence, the sequence
handling was adjusted, and datatype "text" was replaced by
"varchar(4000)" and in one case by "clob".  Probably, some of the
lengths and the update operations have to be adjusted.