Fixes for Oracle 19c: (temporary?) fix for message catalogs
This change fixes the problem that after a fresh installation, all (!)
messages keys are empty. I am not sure what the source of the problem
is, why this has supposedly worked some years before, and whether
there are undesired side-effects of this change. Without this change,
the openacs installation leads to fully broken state.