util::file_content_check: new utility for checking content of a file
Check whether the provided file is of the requested type.
This function is more robust and protable than relying on
external programs and their output, but it does not work on
all possible file types. It checks a few common cases that
could lead to problems otherwise, like when uploading archives.

Bump version to 5.10.1d9