Improve robustness of upgrade script for legacy applications
In case, an installation has the legacy view (must be from very old installations)
in place, it is necessary to drop it before registered_users, otherwise
PostgreSQL will complain that it cannot drop view registered_users because
other objects depend on it.

Many thanks to Franz Penz to report this potential problem for old sites.