Improved toolkit support for homogeneous theming and appearance
This change included new tags for rendering icons and toggles. The
change makes it possible that an unchanged markup of icons can be
adapt itself according to a theme or a manual configuration. One can
use e.g. for Bootstrap 3 glyphicons, and when changing the theme to
Bootstrap 5, then bootstrap icons are used.

Furthermore, the new markup eases maintenance and helps to achieve a
consistent look-and-feel also, when design decisions are altered.

Both, bootstrap-icons and glyphicons are font-based. This leads as
well to a better user experience, since the number of requests are
substantially lower than with the previous approach based on .gif and
.png files. When the iconset is set to "classic" (package parameter in
acs-subsite), the old image files are used.

- new tag: <adp:icon name="NAME" title="..." style="..." class="...">

  Icon-names available for all iconsets:
    edit
    trash
    radio-checked
    radio-unchecked
    checkbox-checked
    checkbox-unchecked
    arrow-down
    arrow-up

  The page acs-admin/package/bootstrap-icons/sample contains more
  detailed information, including the names which are identical
  between bootstrap 3 and 5.


- bumped version number of acs-templating to 5.10.1d6