ad_page/include_contract argspec parsing reform: allow arbitrary characters in the argspec flag parameters
This reform allows to specify an argspec containing pipes, parenthesys and other so far forbidden characters as parameters for a flag. The purpose is to enhance the expressiveness of existing validators (e.g. the oneof validator) and enable new one, for instance, a date validator using a clock format as parameter.
Pipes and parenthesys need to be escaped via the backslash character in the new syntax.