Quick fix for problems with recursive query for type hierarchy on openacs.org
On OpenACS.org, the types for object_type in acs_objects and acs_object_types differs,
leading to a hard error in PostgreSQL. This is just a quick fix, the proper
fix is to investigate, why the types differ (other sites might be as well
affected by this) and to provide a proper update script.

Furthermore, the query is not Oracle compliant.