Adjust legacy data types of acs_object_types.object_type and acs_objects.object_type
This change addresses the problem mentioned in
    https://cvs.openacs.org/changelog/OpenACS?cs=oacs-5-10%3Agustafn%3A20220729185340
    https://github.com/openacs/openacs-core/commit/be44691f06627678122bd913bc3c95c80e93f403

which happens in some legacy applications (such as
e.g. openacs.org) where the data types of the following two
attributes are different.

      acs_object_types.object_type
      acs_objects.object_type

On new installations (at least concerning the last 10 years) these
data types are the same.  So, probably an update script was missing
ages ago.

Since this change affects the basic object structure, MANY views
depend on this datatype and have to be dropped and recreate to
allow the correction of the datatype.

Therefore, we do not want to run this script on all sites, but
only on those where it is necessary.