Generalized "version_dir" handling a little for download specs
The problem was that bootstrap5 uses a version directory, which
consists of the version plus an extra string element. The previous
version assumed that the version is always used as a directory name.

Whe track now the versionDir information in the resource_info dict and
use this, when available (otherwise the version number is used as before).

The resource_info dict contains now the following path components:

 # Provide paths for loading either via /resources/ or CDN
 #
 #   "resourceDir" is the absolute path in the filesystem
 #   "resourceUrl" is the URL path provided to the request processor
 #   "versionDir" is the version-specific element both in the
 #                URL and in the filesystem.
 #

bumped acs-tcl to 5.10.1d19