Changed "db_multirow" to use "db_list_of_lists" instead of "db_list_of_ns_sets"
Thanks to the recent improvements in NaviServer for
"db_list_of_lists", it is more efficient to use this in "db_multirow"
(actually in "db_multirow_helper"). The implementation revealed a bug
in "db_list_of_lists", when both options  "-columns_var" and
"-with_headers" were set.

Extended regression test to cover the former bug.