Rework parameter check:
the way the if was construed would allow e.g.:
 1. a value for EmailRegistrationConfirmationToUserP = 'nottrue' to pass the first if, but fail the second.
 2. a value EmailRegistrationConfirmationToUserP = 'false' to be "true" in the first if, but not in the second

In fact, if EmailRegistrationConfirmationToUserP should be a boolean, there is no need to further check the other parameters and variables in the second if. This change is not 100% equivalent to the previous code, but makes it clear that EmailRegistrationConfirmationToUserP should be interpreted as a boolean and fixes these inconsistencies.