Activate translator mode only when developer support is active
We just do this when the developer support is active, but this does
not have to be this way. By showing the translator mode only for
developer support, we save for large sites many set operations client
for client properties via "lang::util::translator_mode_p" and
"ad_get_client_property".