Remove no_alarms_until column
Rationale: the column or the related api is never used in upstream code to check whether notifications should be sent or not or not. The UI to manage the column's value was not linked anywhere. On the instances accessible by me downstream, no tuple was found where this column had a value.