Test that "apm_get_package_files -all" should return all files for a package
This currently fails, as "-all" does not override the dbms check, so e.g. on a Postgres installation, all Oracle xql files would be skipped.

Note that the test uses acs-tcl as target because it is a package that is only supposed to contain "package-relevant" files. This is not a rule and is not the case in other packages such as acs-automated-testing, which creates many files during its operations that do not belong to the source tree.