bugfix: provide proper encoding for text phrases in email addresses
Previously, mails to addresses of the form

    PHRASE <USER@HOST>

where PHRASE contains UTF-8 were sent without the required encoding.
One many systems, this was perfectly fine, but on some, these
where leading to mangled characters. Now, the "to_addr" adresseses
and the "from_addr" address of acs_mail_lite::send_immediately
are properly encoded. Note, that "to_addr" is a list of addresses.

Many thanks to Franz Penz for reporting this.